factor out into a generate module
